Added posts, post list, likes, comment button, profile icon, and some Display things
We can now make posts all the posts are displayed with relative time since they were posted. You can like and unlike posts click on a profile picture or name to be brought to their profile page. The comment button brings you to a comment page for that post. Trashcan icon there but not functional yet.

Installed react-icons for icons and date-fns for the time each post was posted.