Add text wrap in comment and return for when post delete in comments page
Error when you deleted post while in the comments page fixed now you are
redireted to dashboard. Text will no longer continue to expand box when
comments are made they will only go a set length, it first trys to break
on word space but will break in middle of word if forced to.